ublox-mp8/pyproject.toml
[tool.poetry]
name = "ubloxvenv"
version = "0.1.0"
description = ""
authors = ["Jake Nations <jnations@ucsc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
polars = "^0.17.11"
pynmea2 = "^1.19.0"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
